% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tweet_url_info}
\alias{add_tweet_url_info}
\title{bla}
\usage{
add_tweet_url_info(
  .tweet_tidy_data,
  .url_info_db_path = fs::path(data_dir, "url_info_db.rds"),
  .url_info_db_retry = FALSE
)
}
\arguments{
\item{.tweet_tidy_data}{...}

\item{.url_info_db_path}{...}

\item{.url_info_db_retry}{...}
}
\value{
\code{add_tweet_url_info} - returns a ...
}
\description{
\code{add_tweet_url_info} - Convert all Emojis to some ...
}
\examples{
1+1
}
