% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-init.R
\name{build_nest}
\alias{build_nest}
\title{Initialise}
\usage{
build_nest(edit = interactive())
}
\arguments{
\item{edit}{Whether to open the file for edit.}
}
\description{
Create a \code{_chirp.yml} file in the current working directory.
}
