% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chirp-package.R
\docType{package}
\name{chirp-package}
\alias{chirp}
\alias{chirp-package}
\title{chirp: A collection of tools to collect and process tweets in R}
\description{
...
}
\author{
\strong{Maintainer}: Moritz Pilarski \email{first.last@example.com} (\href{https://orcid.org/0009-0009-5063-0979}{ORCID})

}
\keyword{internal}
