% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_tweet_count_raw}
\alias{fetch_tweet_count_raw}
\title{bla}
\usage{
fetch_tweet_count_raw(
  .search_query,
  .date_new,
  .date_old,
  .resolution,
  .bearer_token,
  .scope = "all",
  ...
)
}
\arguments{
\item{.search_query}{...}

\item{.date_new}{...}

\item{.date_old}{...}

\item{.resolution}{...}

\item{.bearer_token}{...}

\item{...}{...}
}
\value{
\code{fetch_tweet_count_raw} - returns a ...
}
\description{
\code{fetch_tweet_count_raw} - Convert all Emojis to some ...
}
\examples{
1+1
}
