% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_tweet_search_raw}
\alias{fetch_tweet_search_raw}
\title{bla}
\usage{
fetch_tweet_search_raw(.search_query, .date_new, .date_old, .bearer_token, ...)
}
\arguments{
\item{.search_query}{...}

\item{.date_new}{...}

\item{.date_old}{...}

\item{...}{...}
}
\value{
\code{fetch_tweet_search_raw} - returns a ...
}
\description{
\code{fetch_tweet_search_raw} - ...
}
\examples{
1+1
}
