% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_tweet_timeline_raw}
\alias{fetch_tweet_timeline_raw}
\title{bla}
\usage{
fetch_tweet_timeline_raw(
  .user_id_vec = bit64::integer64(),
  .bearer_token,
  .tweet_count_max = Inf,
  .excl_replies = FALSE,
  .excl_retweets = FALSE,
  .until_tweet_id = NULL,
  .page_size = 10L,
  .tweet_query_pars_static = NULL,
  ...
)
}
\arguments{
\item{.user_id_vec}{...}

\item{.bearer_token}{...}

\item{.tweet_count_max}{...}

\item{.excl_replies}{...}

\item{.excl_retweets}{...}

\item{.until_tweet_id}{...}

\item{.page_size}{...}

\item{.tweet_query_pars_static}{...}

\item{...}{...}
}
\value{
\code{fetch_tweet_timeline_raw} - returns a ...
}
\description{
\code{fetch_tweet_timeline_raw} - ...
}
\examples{
1+1
}
