% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-deploy.R
\name{fly}
\alias{fly}
\title{Deploy}
\usage{
fly(edit = interactive())
}
\arguments{
\item{edit}{Whether to open the file for edit.}
}
\description{
Creates the necessary \code{app.R} file in the current working directory.
}
