% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDPMix.r
\name{NDPMix}
\alias{NDPMix}
\title{Function for posterior sampling of DP mixture of Gaussian regressions.}
\usage{
NDPMix(d_train, formula, d_test, burnin, iter, phi_y = c(shape = 5, rate
  = 1000), beta_prior_mean = NULL, beta_prior_var = NULL,
  init_k = 10, beta_var_scale = 10000, x_var_scale = 1,
  b2_scale = 1)
}
\arguments{
\item{d_train}{Input dataset.}
}
\description{
This function takes input dataset and performs posterior sampling. It returns posterior predictive draws and posterior cluster assignments.
}
\keyword{Dirichlet}
