% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data-tweet.R, R/prep_data-tweet_count.R
\name{prep_tidy_data.tweet}
\alias{prep_tidy_data.tweet}
\alias{prep_tidy_data.tweet_count}
\title{bla}
\usage{
prep_tidy_data.tweet(
  .tweet_raw_data,
  .include_media_url = FALSE,
  .get_all = FALSE
)

prep_tidy_data.tweet_count(.tweet_raw_data)
}
\arguments{
\item{.tweet_raw_data}{...}

\item{.include_media_url}{... \code{FALSE}}

\item{.get_all}{output clean data for all included posts (not just directly requested; "main") \code{FALSE}}
}
\value{
\code{prep_tidy_data.tweet} - returns a ...

\code{prep_tidy_data.tweet} - returns a ...
}
\description{
\code{prep_tidy_data.tweet} - Convert all Emojis to some ...

\code{prep_tidy_data.tweet} - Convert all Emojis to some ...
}
\examples{
1+1
1+1
}
