% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_tweet_data_tidy}
\alias{prep_tweet_data_tidy}
\title{bla}
\usage{
prep_tweet_data_tidy(.tweet_data_raw, .include_media_url = FALSE)
}
\arguments{
\item{.tweet_data_raw}{...}

\item{.include_media_url}{... \code{FALSE}}
}
\value{
\code{prep_tweet_data_tidy} - returns a ...
}
\description{
\code{prep_tweet_data_tidy} - Convert all Emojis to some ...
}
\examples{
1+1
}
