% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_pars_static_dict}
\alias{query_pars_static_dict}
\title{Static Query Parameters}
\format{
An object of class \code{list} of length 2.
}
\usage{
query_pars_static_dict
}
\description{
Static Query Parameters
}
\references{
\url{https://docs.x.com/x-api}
}
\keyword{datasets}
