% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_tweet_is_convers_start}
\alias{add_tweet_is_convers_start}
\title{bla}
\usage{
add_tweet_is_convers_start(.tweet_data_tidy, .na_to_false = FALSE)
}
\arguments{
\item{.tweet_data_tidy}{...}

\item{.na_to_false}{... \code{FALSE}}
}
\value{
\code{add_tweet_is_convers_start} - returns a ...
}
\description{
\code{add_tweet_is_convers_start} - Convert all Emojis to some ...
}
\examples{
1+1
}
