% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_assign_mode.r
\name{cluster_assign_mode}
\alias{cluster_assign_mode}
\title{Function for computing the posterior mode of cluster assignment.}
\usage{
cluster_assign_mode(c_shell)
}
\arguments{
\item{d_train}{Input dataset.}
}
\description{
This function corrects for label switching and outputs posterior mode cluster assignment. Label switching correction is deterministic - finding the posterior adjacency matrix that is closest to the posterior mode adjacency matrix in the L2 sense.
}
\keyword{Label}
\keyword{Switching}
