% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_tweet_id_raw}
\alias{fetch_tweet_id_raw}
\alias{fetch_tweet_search_raw}
\alias{fetch_tweet_count_raw}
\title{bla}
\usage{
fetch_tweet_id_raw(
  .tweet_id_vec,
  .bearer_token,
  .tweet_query_pars_static = NULL,
  ...
)

fetch_tweet_search_raw(.search_query, .date_new, .date_old, .bearer_token, ...)

fetch_tweet_count_raw(
  .search_query,
  .date_new,
  .date_old,
  .resolution,
  .bearer_token,
  ...
)
}
\arguments{
\item{.tweet_id_vec}{...}

\item{.bearer_token}{...}

\item{.tweet_query_pars_static}{...}

\item{...}{...}

\item{.search_query}{...}

\item{.date_new}{...}

\item{.date_old}{...}

\item{.resolution}{...}
}
\value{
\code{fetch_tweet_id_raw} - returns a ...

\code{fetch_tweet_id_raw} - returns a ...

\code{fetch_tweet_id_raw} - returns a ...
}
\description{
\code{fetch_tweet_id_raw} - Convert all Emojis to some ...

\code{fetch_tweet_search_raw} - Convert all Emojis to some ...

\code{fetch_tweet_search_raw} - Convert all Emojis to some ...
}
\examples{
1+1
1+1
1+1
}
