% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDPMix.r
\name{PDPMix}
\alias{PDPMix}
\title{Function for posterior sampling of DP mixture of zero-inflated regressions.}
\usage{
PDPMix(d_train, formula, d_test, burnin, iter, beta_prior_mean = NULL,
  beta_prior_var = NULL, prop_sigma_b = diag(rep(0.025, nparams)),
  init_k = 10, beta_var_scale = 1000, mu_scale = 1, tau_scale = 1)
}
\arguments{
\item{d_train}{Input dataset.}
}
\description{
This function takes input dataset and performs posterior sampling. It returns posterior predictive draws, posterior clustering, and posterior parameter draws.
}
\keyword{Dirichlet}
