% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{query_pars_static_dict}
\alias{query_pars_static_dict}
\title{Static Query Parameters}
\format{
\subsection{\code{query_pars_static_dict}}{

A list:
\describe{
\item{names}{...}
\item{values}{...}
...
}
}
}
\source{
\url{https://docs.x.com/x-api}
}
\usage{
query_pars_static_dict
}
\description{
...
}
\keyword{datasets}
